feat: Added some chart logic and changed schema little bit

+++ src/lib/actions/UserHealthActions.tsx
@@ -2,16 +2,18 @@
 
 import { db } from "@/db";
 import { userHealthInfo } from "@/db/schema";
-import { eq } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 import { z } from "zod";
 import { getCurrentSession } from "../sessionTokens";
 
 const createFormSchema = z.object({
-  weight: z.number().optional(),
-  height: z.number().optional(),
-  age: z.number().optional(),
-  gender: z.enum(["male", "female"]).optional(),
+  weight: z.coerce.number().nonnegative(),
+  weightPounds: z.coerce.boolean(),
+  height: z.coerce.number().nonnegative(),
+  heightInches: z.coerce.boolean(),
+  age: z.coerce.number().nonnegative().max(120),
+  gender: z.enum(["male", "female"]),
   activityLevel: z
     .enum([
       "sedentary",
@@ -19,13 +21,24 @@ const createFormSchema = z.object({
       "moderately_active",
       "very_active",
       "extremely_active",
+    ]),
+  calories: z.coerce.number().nonnegative(),
+  fitnessGoal: z
+    .enum([
+      "lose_weight",
+      "gain_muscle",
+      "maintain_weight",
     ])
-    .optional(),
-  calories: z.number().optional(),
 });
+
 
 export async function createUserHealth(formData: FormData) {
   try {
+    const { user, session } = await getCurrentSession();
+    if (!user || !session) {
+      return { error: "Unauthorized", code: 401 };
+    }
+
     const parsedFormData = Object.fromEntries(formData.entries());
 
     const validatedData = createFormSchema.safeParse(parsedFormData);
@@ -33,14 +46,11 @@ export async function createUserHealth(formData: FormData) {
       return { error: "Invalid data", code: 400 };
     }
 
-    const { weight, height, age, gender, activityLevel, calories } =
+    const { fitnessGoal, weight, height, age, gender, activityLevel, calories, weightPounds, heightInches } =
       validatedData.data;
-    const { user } = await getCurrentSession();
-    if (!user) {
-      return { error: "Invalid data", code: 400 };
-    }
 
     await db.insert(userHealthInfo).values({
+      fitnessGoal: fitnessGoal,
       userId: user.id,
       weight: weight,
       height: height,
@@ -48,16 +58,25 @@ export async function createUserHealth(formData: FormData) {
       gender: gender,
       activityLevel: activityLevel,
       calories: calories,
+      weightPounds: weightPounds,
+      heightInches: heightInches,
     });
 
     revalidatePath("/main/dashboard");
   } catch (e) {
+    console.log(e);
     return { error: "Server error", code: 500 };
   }
 }
 
 export async function updateUserHealth(formData: FormData, id: string) {
   try {
+
+    const { user, session } = await getCurrentSession();
+    if (!user || !session) {
+      return { error: "Unauthorized", code: 401 };
+    }
+
     const parsedFormData = Object.fromEntries(formData.entries());
     const id_val = z.string().max(36).safeParse(id);
 
@@ -70,12 +89,14 @@ export async function updateUserHealth(formData: FormData, id: string) {
       return { error: "Invalid data", code: 400 };
     }
 
-    const { weight, height, age, gender, activityLevel, calories } =
+    const { weight, height, age, gender, activityLevel, calories, weightPounds, heightInches } =
       validatedData.data;
 
     await db
       .update(userHealthInfo)
       .set({
+        weightPounds: weightPounds,
+        heightInches: heightInches,
         weight: weight,
         height: height,
         age: age,
@@ -83,7 +104,7 @@ export async function updateUserHealth(formData: FormData, id: string) {
         activityLevel: activityLevel,
         calories: calories,
       })
-      .where(eq(userHealthInfo.id, id_val.data));
+      .where(and(eq(userHealthInfo.id, id_val.data), eq(userHealthInfo.userId, user.id)));
 
     revalidatePath("/main/dashboard");
   } catch (e) {
@@ -93,13 +114,20 @@ export async function updateUserHealth(formData: FormData, id: string) {
 
 export async function deleteUserHealth(id: string) {
   try {
+
+    const { user, session } = await getCurrentSession();
+    if (!user || !session) {
+      return { error: "Unauthorized", code: 401 };
+    }
+
+
     const id_val = z.string().max(36).safeParse(id);
 
     if (!id_val.success) {
       return { error: "Invalid data", code: 400 };
     }
 
-    await db.delete(userHealthInfo).where(eq(userHealthInfo.id, id));
+    await db.delete(userHealthInfo).where(and(eq(userHealthInfo.id, id), eq(userHealthInfo.userId, user.id)));
     revalidatePath("/main/dashboard");
   } catch (e) {
     return { error: "Server error", code: 500 };
@@ -108,6 +136,11 @@ export async function deleteUserHealth(id: string) {
 
 export async function getUserHealth(id: string) {
   try {
+    const { user, session } = await getCurrentSession();
+    if (!user || !session) {
+      return { error: "Unauthorized", code: 401 };
+    }
+
     const id_val = z.string().max(36).safeParse(id);
 
     if (!id_val.success) {
@@ -117,10 +150,11 @@ export async function getUserHealth(id: string) {
     const user_health = await db
       .select()
       .from(userHealthInfo)
-      .where(eq(userHealthInfo.id, id_val.data));
+      .where(eq(userHealthInfo.userId, user.id));
 
     return user_health[0] ?? null;
   } catch (e) {
+    console.log(e);
     return { error: "Server error", code: 500 };
   }
 }
